Refactoring

+++ src/util/data/storage/Session.java
@@ -11,7 +11,12 @@ import java.util.Map;
 
 /**
  * Session manager that stores information obtained at run time. It allows 
- * different processes and threads to have access to this information.
+ * different processes and threads to have access to this information. There
+ * are two types of session:
+ * <ul>
+ * 	<li>Normal - disk storage</li>
+ * 	<li>Shared - storage in class data memory</li> 
+ * </ul>
  * 
  * @author		William Niemiec &lt; williamniemiec@hotmail.com &gt;
  */
@@ -33,8 +38,17 @@ public class Session {
 	 * 
 	 * @param		name Session name
 	 * @param		directory Location where the session will be saved
+	 * 
+	 * @throws		IllegalArgumentException If name is null or empty or if
+	 * directory is null
 	 */
 	public Session(String name, File directory)	{
+		if ((name == null) || name.isBlank())
+			throw new IllegalArgumentException("Name cannot be empty");
+		
+		if (directory == null)
+			throw new IllegalArgumentException("Directory cannot be null");
+		
 		this.sessionFile = new File(directory, name + ".bin");
 		this.content = new HashMap<>();
 	}
@@ -52,9 +66,17 @@ public class Session {
 	 * @throws		IOException If an error occurred while storing the session
 	 * 
 	 * @implNote	To allow different processes and threads to have access to 
-	 * session data, it is stored on disk.
+	 * session data, it is stored on disk
+	 * 
+	 * @throws		IllegalArgumentException If key or value is null
 	 */
 	public synchronized void save(String key, Object value) throws IOException {
+		if (key == null)
+			throw new IllegalArgumentException("Key cannot be null");
+		
+		if (value == null)
+			throw new IllegalArgumentException("Value cannot be null");
+		
 		if (exists()) {
 			load();
 			sessionFile.delete();			
@@ -64,15 +86,7 @@ public class Session {
 		
 		content.put(key, value);
 		
-		try (ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(sessionFile))) {
-			oos.writeObject(content);
-			oos.flush();
-		}
-	}
-	
-	public static synchronized void saveShared(String key, Object value) 
-			throws IOException {
-		sharedContent.put(key, value);
+		store();
 	}
 
 	/**
@@ -83,11 +97,17 @@ public class Session {
 	 * @return		Stored information of null if there is no data stored with 
 	 * the provided identifier
 	 * 
+	 * @implNote	It is read from class data memory
+	 * 
 	 * @throws		IOException If an error occurred while loading the session
 	 * @throws		IllegalStateException If session does not exist. For a 
 	 * session to exist, something must be saved in it
+	 * @throws		IllegalArgumentException If key is null
 	 */
 	public Object read(String key) throws IOException {
+		if (key == null)
+			throw new IllegalArgumentException("Key cannot be null");
+		
 		if (!exists())
 			throw new IllegalStateException("Session does not exist");
 		
@@ -96,11 +116,31 @@ public class Session {
 		return content.get(key);
 	}
 	
-	public static Object readShared(String key) {
-		if (!sharedContent.containsKey(key))
-			throw new IllegalStateException("Key not found");
+	/**
+	 * Removes data from the current session.
+	 * 
+	 * @param		key Information identifier to be removed
+	 * 
+	 * @throws		IOException If an error occurred while loading the session
+	 * @throws		IllegalStateException If session does not exist. For a 
+	 * session to exist, something must be saved in it
+	 * @throws		IllegalArgumentException If key is null
+	 */
+	public void remove(String key) throws IOException {
+		if (key == null)
+			throw new IllegalArgumentException("Key cannot be null");
 		
-		return sharedContent.get(key);
+		if (!exists())
+			throw new IllegalStateException("Session does not exist");
+		
+		load();
+		
+		if (!content.containsKey(key))
+			return;
+		
+		content.remove(key);
+		
+		store();
 	}
 	
 	/**
@@ -108,21 +148,20 @@ public class Session {
 	 * 
 	 * @param		key Identifier
 	 * 
-	 * @return		True if the key is associated with any data.
+	 * @return		True if the key is associated with any data
+	 * 
+	 * @throws		IllegalArgumentException If key is null
+	 * @throws		IllegalStateException If session does not exist. For a 
+	 * session to exist, something must be saved in it
 	 */
 	public boolean hasKey(String key) {
-		return content.containsKey(key);
-	}
-	
-	public static boolean hasKeyShared(String key) {
-		return sharedContent.containsKey(key);
-	}
-	
-	public static void removeShared(String key) {
-		if (!sharedContent.containsKey(key))
-			return;
+		if (key == null)
+			throw new IllegalArgumentException("Key cannot be null");
 		
-		sharedContent.remove(key);
+		if (!exists())
+			throw new IllegalStateException("Session does not exist");
+		
+		return content.containsKey(key);
 	}
 	
 	/**
@@ -162,5 +201,97 @@ public class Session {
 		}		
 		catch (ClassNotFoundException e) {
 		}
+	}
+	
+	private void store() throws IOException {
+		try (ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(sessionFile))) {
+			oos.writeObject(content);
+			oos.flush();
+		}
+	}
+	
+	/**
+	 * Stores data in the current session.
+	 * 
+	 * @param		key Identifier so that it is possible to retrieve data
+	 * @param		value Data to be stored
+	 * 
+	 * @throws		IOException If an error occurred while storing the session
+	 * 
+	 * @implNote	To allow different processes and threads to have access to 
+	 * session data, it is stored in class data memory
+	 * 
+	 * @throws		IllegalArgumentException If key or value is null
+	 */
+	public static synchronized void saveShared(String key, Object value) 
+			throws IOException {
+		if (key == null)
+			throw new IllegalArgumentException("Key cannot be null");
+		
+		if (value == null)
+			throw new IllegalArgumentException("Value cannot be null");
+		
+		sharedContent.put(key, value);
+	}
+	
+	/**
+	 * Gets data from the current session.
+	 * 
+	 * @param		key Information identifier 
+	 * 
+	 * @return		Stored information of null if there is no data stored with 
+	 * the provided identifier
+	 * 
+	 * @implNote	It is read from class data memory
+	 * 
+	 * @throws		IOException If an error occurred while loading the session
+	 * @throws		IllegalStateException If session does not exist. For a 
+	 * session to exist, something must be saved in it
+	 * @throws		IllegalArgumentException If key is null
+	 */
+	public static Object readShared(String key) {
+		if (key == null)
+			throw new IllegalArgumentException("Key cannot be null");
+		
+		if (!sharedContent.containsKey(key))
+			throw new IllegalStateException("Key not found");
+		
+		return sharedContent.get(key);
+	}
+	
+	/**
+	 * Checks  if there is data stored in the session with the specified key.
+	 * 
+	 * @param		key Identifier
+	 * 
+	 * @return		True if the key is associated with any data
+	 * 
+	 * @throws		IllegalArgumentException If key is null
+	 */
+	public static boolean hasKeyShared(String key) {
+		if (key == null)
+			throw new IllegalArgumentException("Key cannot be null");
+		
+		return sharedContent.containsKey(key);
+	}
+	
+	/**
+	 * Removes data from the current session.
+	 * 
+	 * @param		key Information identifier to be removed
+	 * 
+	 * @throws		IOException If an error occurred while loading the session
+	 * @throws		IllegalStateException If session does not exist. For a 
+	 * session to exist, something must be saved in it
+	 * @throws		IllegalArgumentException If key is null
+	 */
+	public static void removeShared(String key) {
+		if (key == null)
+			throw new IllegalArgumentException("Key cannot be null");
+
+		if (!sharedContent.containsKey(key))
+			return;
+		
+		sharedContent.remove(key);
 	}
 }
